Move SaveDailyMemoResult to domain/model
For consistency with other domain types, move SaveDailyMemoResult
sealed interface from inline definition in SaveDailyHabitMemoUseCase
to its own file in domain/model package.
